Use normalized fix_type for HoTT GPS fix character
Per review feedback on this PR, the previous approach (remapping the
raw GPS fix value locally inside hott.c) duplicates the raw->simplified
mapping that gps.c already does in `parameter.fix_type`. The correct
fix is to wire HoTT's HOTT_GPS_FIX off the normalized value, keep
`parameter.fix` as the raw exposure for callers that want all UBLOX
states (DR, GNSS+DR, time-only), and remove the local remap in hott.c.

To pass the normalized value through the existing `sensors->gps[]`
array, which holds `float *`, change the type of `parameter.fix_type`
from `uint8_t *` to `float *`. The float can store 0/1/2 just as well,
existing assignments auto-convert from int literals, and the protocols
that read `*fix_type` only use comparisons or casts.

Touched in lock-step:
- gps.h, distance.h: fix_type / distance fix to `float *`
- 12 protocol files (all rx-protocol setups): allocate fix_type with
  `sizeof(float)` instead of `sizeof(uint8_t)`
- ghst.c, xbus.h struct fix_type pointer fields: now `float *`
- sbus.c local `gps_fix` cache pointer: now `float *`
- hott.c: wire `HOTT_GPS_FIX = parameter.fix_type`; interpretation
  in `format_binary_packet` stays as `0 -> ' ', 1 -> '2', else -> '3'`,
  which is now correct because the value is normalized

Side effect: the `gnssFixOK` consistency case (UBX `fixType` reports
2D/3D but flag says low confidence, e.g. PDOP ~50 during acquisition,
visible in @slowake's Test1 log on #194) is now handled correctly --
gps.c already resets fix_type to 0 when `!(flags & 0x01)`, so HoTT
will now show blank instead of a misleading "2D" during the
acquisition transition.

I now have a BN-220 module on hand and plan to bench-verify this
against a real fix as soon as the setup is wired up.

Addresses item (2) of #194. Built with Pico SDK 2.2.0 / ARM GCC 15.2.

+++ board/project/protocol/ghst.c
@@ -99,7 +99,7 @@ typedef struct ghst_sensor_gps_secondary_t {
     float *homedir;
     float *flags;
     uint8_t *home_set;
-    uint8_t *fix_type;
+    float *fix_type;
 } ghst_sensor_gps_secondary_t;
 
 typedef struct ghst_sensor_magbaro_t {
@@ -300,7 +300,7 @@ static void set_config(ghst_sensors_t *sensors) {
         parameter.v_vel = malloc(sizeof(float));
         parameter.alt_elipsiod = malloc(sizeof(float));
         parameter.pdop = malloc(sizeof(float));
-        parameter.fix_type = malloc(sizeof(uint8_t));
+        parameter.fix_type = malloc(sizeof(float));
         parameter.home_set = malloc(sizeof(uint8_t));
         parameter.alt_home = malloc(sizeof(float));
         xTaskCreate(gps_task, "gps_task", STACK_GPS, (void *)&parameter, 2, &task_handle);
