+++ board/project/protocol/sbus.c
@@ -114,7 +114,7 @@ typedef struct sensor_sbus_t {
 } sensor_sbus_t;
 
 static sensor_sbus_t *sbus_sensor[32] = {NULL};
-static uint8_t *gps_fix;
+static float *gps_fix;
 static uint packet_id;
 static volatile uint slot = 0;
 static volatile bool slot_pending = false;
@@ -721,7 +721,7 @@ static void set_config(void) {
         parameter.v_vel = malloc(sizeof(float));
         parameter.alt_elipsiod = malloc(sizeof(float));
         parameter.pdop = malloc(sizeof(float));
-        parameter.fix_type = malloc(sizeof(uint8_t));
+        parameter.fix_type = malloc(sizeof(float));
         parameter.home_set = malloc(sizeof(uint8_t));
         parameter.alt_home = malloc(sizeof(float));
         xTaskCreate(gps_task, "gps_task", STACK_GPS, (void *)&parameter, 2, &task_handle);
